+++ src/Akeneo/AssetManager/front/application/component/asset/create.tsx
@@ -8,7 +8,7 @@ import {createLocaleFromCode, LocaleCode} from 'akeneoassetmanager/domain/model/
 import Checkbox from 'akeneoassetmanager/application/component/app/checkbox';
 import {Asset} from 'akeneoassetmanager/application/component/app/illustration/asset';
 import AssetCode from 'akeneoassetmanager/domain/model/asset/code';
-import sanitizeAssetCode from 'akeneoassetmanager/tools/sanitizeAssetCode';
+import {sanitizeAssetCode} from 'akeneoassetmanager/tools/sanitizeAssetCode';
 import AssetFamilyIdentifier from 'akeneoassetmanager/domain/model/asset-family/identifier';
 import assetSaver from 'akeneoassetmanager/infrastructure/saver/asset';
 import {useFocus, useShortcut} from 'akeneoassetmanager/application/hooks/input';