+++ CHANGELOG.md
@@ -46,6 +46,7 @@
 - PIM-9500bis: Add "asset_manager_link_assets_to_products" and "asset_manager_compute_transformations" type and job translations.
 - PIM-9529: Fix translation in rule engine edit page
 - PIM-9536: Fix unexpected behaviors on drag&drop in rules edit page (calculate and concatenate actions)
+- PIM-9528: Fix asset code changed into lower case in create asset/upload asset UI
 
 ## Improvements